Fix async unit test failure, update readme links
